requirements.txt
fastapi
uvicorn
python-dotenv
requests
llama-index-core
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-vector-stores-elasticsearch
sentence-transformers
protobuf
python-multipart
pymongo
elasticsearch